feat: Add DB Connect Code

+++ omok/src/main/java/VO/UserVO.java
@@ -6,9 +6,9 @@ import lombok.Setter;
 @Getter
 @Setter
 public class UserVO {
-    private Long userId;
+    private int userId;
     private String userName;
     private String userPw;
-    private Long userWinCnt;
-    private Long userGameCnt;
+    private int userWinCnt;
+    private int userGameCnt;
 }
